Add User, Thoughts, Reactions schemas and setup routing
Added schema models for User, Thoughts, and Reactions with accompanying fields, validations, and virtual fields for counts. Re-named and moved a file from controllers to models. Also, set up root level routing in 'routes/index.js' with an endpoint for API routes and a status 404 default response for undefined routes.